backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
pydantic
python-multipart
python-dateutil
python-jose[cryptography]
passlib[bcrypt]
